fix: fix macos code checker (#36817)
https://github.com/milvus-io/milvus/issues/36829

---------

+++ internal/core/src/mmap/ChunkedColumn.h
@@ -123,7 +123,7 @@ class ChunkedColumnBase : public ColumnBase {
         chunks_.push_back(chunk);
     }
 
-    virtual uint64_t
+    virtual size_t
     DataByteSize() const override {
         auto size = 0;
         for (auto& chunk : chunks_) {
@@ -353,8 +353,8 @@ class ChunkedVariableColumn : public ChunkedColumnBase {
                 {chunks_[start_chunk_id]->Data(),
                  std::dynamic_pointer_cast<StringChunk>(chunks_[start_chunk_id])
                      ->Offsets(),
-                 start_offset_in_chunk,
-                 start_offset_in_chunk + len});
+                 static_cast<int>(start_offset_in_chunk),
+                 static_cast<int>(start_offset_in_chunk + len)});
 
             start_offset_in_chunk = 0;
             length -= len;
